% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterParams.R
\name{clusterParams}
\alias{clusterParams}
\title{clusterParams}
\usage{
clusterParams(data)
}
\arguments{
\item{data}{\code{dataframe} -- dataframe with cluster labels}
}
\value{
A \code{list} of 2 elements:
\describe{
\item{ClusterParams}{\code{numeric} -- the estimated cluster parameters.}
\item{mixtureWeights}{\code{numeric} -- the estimated mixture weights for the clusters.}
}
}
\description{
Title
}
\examples{

}
