% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPrediction.R
\name{clusterPredictions}
\alias{clusterPredictions}
\title{Title}
\usage{
clusterPredictions(
  dataWithLabels,
  dataWithoutLabels,
  trainingParams,
  trainMixWeights
)
}
\arguments{
\item{dataWithLabels}{-- data with cluster labels (ideally obtained from PReMiuM)}

\item{dataWithoutLabels}{-- data for which you want to predict cluster labels}

\item{trainingParams}{-- cluster parameters obtained from training set}

\item{trainMixWeights}{-- mixture weights obtained from training set}
}
\value{
maximum likelihood predictions for data without labels
}
\description{
Title
}
