% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{combineForeachOutput}
\alias{combineForeachOutput}
\title{The combineForeachOutput function}
\usage{
combineForeachOutput(a, b)
}
\arguments{
\item{a}{output from one foreach loop}

\item{b}{output from another foreach loop}
}
\description{
Internal. Function to combine output from foreach iterations.
}
\keyword{internal}

