% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{connectivityVector}
\alias{connectivityVector}
\title{The connectivityVector function}
\usage{
connectivityVector(sampleID, groupID, connectVector, n)
}
\arguments{
\item{sampleID}{vector indicating which are the samples being considered}

\item{groupID}{vector with group indexes. Positions corresponding to pair of samples present in sampleID with the same groupID will be increased in 1.}

\item{connectVector}{Current vector of counts to be updated}

\item{n}{total number of samples}
}
\value{
Updated vector of counts
}
\description{
Internal. Given a vector indicating the samples involved (samplesID) and a vector indicating to which group they belong (groupID), 
this function adds 1 to the positions in connectVector corresponding to pair of samples that belong to the same group.
We use this to: 1) update jointOccurrenceVector adding 1 to the positions corresponding to each pair of samples present in the current bootstrap subsample, to indicate that they have the chance to be clustered together later
2) update consensusVector adding 1 to the positions corresponding to pairs of samples that were clustered together
}
\keyword{internal}

