% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{getFinalPartition}
\alias{getFinalPartition}
\title{The getFinalPartition function}
\usage{
getFinalPartition(consensusVector, K, n, sampleNames, finalLinkage, pathOutput,
  verbose)
}
\arguments{
\item{consensusVector}{list a vector of counts of how many times each pair of samples have been clustered together so far, for each value of K.}

\item{K}{vector of integers representing numbeer of clusters to evaluate.}

\item{n}{number of samples in input data matrix}

\item{sampleNames}{vector with samples names (colnames of input data)}

\item{finalLinkage}{heirarchical linkage method for producing a final classification with the consensus indexes}

\item{pathOutput}{directory for output files}

\item{verbose}{logical, print progress messages to screen}
}
\value{
A list with one element for each value of K, each containing:
\itemize{
 \item \code{consensusTree}: final heirarchical tree based on the matrix of consensus indexes after running all the iterations of the clustering.
 \item \code{consensusClass}: vector with the final cluster assignment for each sample.
 \item \code{consensusVector}: vector of consensus indexes for each pair of samples. The consensus index is the proportion of times that a pair of samples 
 was clustered together in the same group, out of the total number times they were in the same bootstrap sample.
}
}
\description{
Internal. Takes the consensus vector for each K and performs hierachical clustering to get the final k clusters.
}
\keyword{internal}

