% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{imputeSampleNumbers}
\alias{imputeSampleNumbers}
\title{Title: Imput Sample Numbers}
\usage{
imputeSampleNumbers(dataframe, numberOfSamples)
}
\arguments{
\item{dataframe}{-- dataframe upon which imputation of sample numbers is to be performed}

\item{numberOfSamples}{-- number of subsamples to be included in data}
}
\value{
dataframe with different datapoints assigned subsample labels
}
\description{
\description{Inorder to carry out the divide-and-conquer consensus clustering, we need to
split a dataset into multiple subsamples. This function assigns points in
the dataset to different subsamples randomly.}
}
\examples{
df <- imputeSampleNumbers(dataframe = iris, numberOfSamples = 3)
}
