% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{lebel_similarity}
\alias{lebel_similarity}
\title{Similarity between different clusters}
\usage{
lebel_similarity(x1, x2)
}
\arguments{
\item{x1}{clustering vector 1
Zero elements are are considered as unclustered samples}

\item{x2}{clustering vector 2
Zero elements are are considered as unclustered samples}
}
\value{
matrix of similarities between clustering labels
}
\description{
Similarity between different clusters
}
\details{
When performing several clustering, the cluster labels may not match with each other.
To find correspondences between clusters, the similarity between different labels will be calculated.
}
\examples{
X = gaussian_clusters()$X
x1 = kmeans(X, 5)$cluster
x2 = kmeans(X, 5)$cluster
Sim = lebel_similarity(x1, x2)

}
