% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Logit}
\alias{Logit}
\title{Logit function}
\usage{
Logit(x)
}
\arguments{
\item{x}{numerical scaler input}
}
\value{
Logit(x) = log(1*x/(1-x))
}
\description{
Logit function
}
\examples{
y = Logit(0.5)

}
