% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{PlotHeatmaps}
\alias{PlotHeatmaps}
\title{The PlotHeatmaps function}
\usage{
PlotHeatmaps(results, plotHeatmaps = c("both", "consensus", "data"),
  plotSave = c("no", "pdf", "bmp", "png", "ps"), pathOutput = "",
  showDendrogram = TRUE, showSamplesNames = TRUE,
  showFeaturesNames = TRUE, dataMatrix = NULL)
}
\arguments{
\item{results}{output from consensusClustering function.}

\item{plotHeatmaps}{character string indicating which heatmaps should be produced: "consensus" (only heatmap of the consensus indexes), "data" (only heatmap of input data set), or "both" (default).}

\item{plotSave}{character string indicating the format the plot to be saved as files in directory \code{pathOutput}. Default is "no", the plots are not saved, but printed to the screen. Other options are: "pdf", "bmp", "png", "ps".}

\item{pathOutput}{directory for plot files}

\item{showDendrogram}{logical indicating if dendrograms should be plotted in the heatmaps (defaults to TRUE).}

\item{showSamplesNames}{logical indicating if sample names should be displayed in the plots (defaults to TRUE).}

\item{showFeaturesNames}{logical indicating if features names should be displayed in the plots (defaults to TRUE).}

\item{dataMatrix}{matrix with data to cluster, samples/items in the columns and features in the rows, only needed if plotHeatmaps is "data" or "both".}
}
\description{
Plots heatmaps of the consensus indexes and original data matrix.
}
\details{
The consensus heatmap displays the matrix of consensus indexes for all pairs of samples. The consensus index is the proportion of times that a pair of samples was clustered together in the same group, out of the total number times they were in the same bootstrap sample. We expect to observe high consensus indexes for samples that were clustered together and low for samples in finally assigned in different clusters. The Consensus heatmap is a visual aid allowing for the consensus clustering membership to be depicted in an easy to understand approach. Cluster membership are highlighted by the colored rectangles between the dendrogram (top) and heatmap according to the legend within the graphic. This layout is beneficial to the user to help compare a clusters' membership count in the context of their consensus. Ideally we aim for high consensus in the blocks along the diagonal (samples in the same cluster) and low consensus outside the diagonal.
}
\examples{
mat <- matrix(rnorm(10*6), 10, 6)
result <- consensusClustering(mat, K=2:3, nIters = 5, plotHeatmaps = "no")
PlotHeatmaps(result, plotHeatmaps = "both", dataMatrix = mat)

}
\keyword{external}

