% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{runForeachLoop}
\alias{runForeachLoop}
\title{The runForeachLoop function}
\usage{
runForeachLoop(nIters, propSamples, dataMatrix, K, clusterAlgorithm, verbose,
  seed)
}
\arguments{
\item{nIters}{number of iterations (bootstrap samples).}

\item{propSamples}{proportion of items to sample in each bootstrap sample.}

\item{dataMatrix}{matrix or data frame with data to cluster, samples/items in the columns and features in the rows.}

\item{K}{vector of integers representing numbeer of clusters to evaluate. It can be of length 1 and it does not need to consist of consecutive integers. For example, either of \code{K = 4}, \code{K = 2:5} or \code{K = c(5, 10, 15)} would work.}

\item{clusterAlgorithm}{algorithm to perform the clustering, for the moment only K-means is available.}

\item{verbose}{logical, print progress messages to screen. During the bootstrap iterations, a report to monitor the progress is created in \code{pathOutput}.}

\item{seed}{numerical value to set random seed for reproducible results. It uses \code{doRNG} package to guarantee reproducible results even when running in parallel.}
}
\value{
A list with the jointOccurrenceVector and consensusVector.
}
\description{
Internal. Wrapper function to call foreach, to keep the code tidy.
}
\keyword{internal}

