% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{selectSubsample}
\alias{selectSubsample}
\title{The selectSubsample function}
\usage{
selectSubsample(nSamples, propSamples)
}
\arguments{
\item{nSamples}{number of total items}

\item{propSamples}{proportion of items to be sampled}
}
\value{
A list in with a vector with the indices of selected samples.
}
\description{
Internal. Select the samples for each bootstrap iteration.
Although it is only one line, we put this task in its own function to easily add more functionality in the future, like also subsetting the features, or giving different weights for the samples
}
\keyword{internal}

