% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{setClusterColors}
\alias{setClusterColors}
\title{The setClusterColors function}
\usage{
setClusterColors(previousConsensusClass, consensusClass, listColors,
  colorTracking, K)
}
\arguments{
\item{previousConsensusClass}{vector of cluster classifications for the previous value of K.}

\item{consensusClass}{vector of cluster classifications for the current value of K.}

\item{listColors}{list of colors to be used.}

\item{colorTracking}{list for storing the colors assigned for the current value of K.}

\item{K}{vector of integers representing numbeer of clusters to evaluate, minimum 2.}
}
\value{
A list with colors assignments for current value of K.
}
\description{
Internal. Sets common color of clusters between different K. Taken from the package ConsensusClusteringPlus, we only modified some variable names.
}
\note{
Function taken from ConsensusClusterPlus package.
}
\references{
Wilkerson M and Hayes D (2010) ConsensusClusterPlus: a class discovery tool with confidence assessments and item tracking. \emph{Bioinformatics}, \strong{26}, 1572-1573.
}
\keyword{internal}

