% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{trackColors}
\alias{trackColors}
\title{The trackColors function}
\usage{
trackColors(results)
}
\arguments{
\item{results}{output from getFinalPartition function.}
}
\value{
A list with the colors assigned to each cluster and sample for each value of K. This colors are assigned tracking the samples trying to preserve same of groups of samples with the same color through different values of K.
}
\description{
Internal. Wrapper function to call function \code{\link{setClusterColors}} from \code{ConsensusClusterPlus} package. Does color assignment to clusters and samples tracking samples across different values of K.
}
\keyword{internal}

