% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusClusteringCode.R
\name{vectToMatr}
\alias{vectToMatr}
\title{The vectToMatr function}
\usage{
vectToMatr(v, N)
}
\arguments{
\item{v}{input vector.}

\item{N}{number of columns in output matrix.}
}
\value{
A symmetric matrix.
}
\description{
Internal. Takes a vector and converts it into a symmetric matrix.
}
\keyword{internal}

